In the Purchase Summary line of the Mall Search Frame, correctly pluralize the item being bought and correctly name (and pluralize) the currency being spent.
git-svn-id: http://svn.code.sf.net/p/kolmafia/code@15070 b29ace70-8910-0410-8dcc-aa2fc6433167

+++ src/net/sourceforge/kolmafia/request/CoinMasterPurchaseRequest.java
@@ -91,7 +91,7 @@ public class CoinMasterPurchaseRequest
 	}
 
 	@Override
-	public String getCurrency( final int count )
+	public String getCurrency( final long count )
 	{
 		return this.cost.getPluralName( this.price );
 	}
